requirements.txt
panda
streamlit
pandasai
pygwalker